requirements.txt
pycozmo
requests
Pillow
numpy
TTS
